Move all saucelabs scripts into a properly named folder
