+++ scripts/saucelabs/list-supported-browsers.js
@@ -2,7 +2,7 @@
 
 var username = 'storejs'
 var password = new Buffer('ZjhjMzUyNjgtNzc2ZC00ZjlkLWEwNWUtN2FkM2Q0ZDgyNzk5', 'base64').toString('utf8')
-var saucelabs = require('./util/saucelabs')
+var saucelabs = require('./saucelabs')
 
 saucelabs.setAuth(username, password)
 saucelabs.listAllSupportedPlatforms(function(err, res) {
